Refactor: practice code in separated files
