+++ java/src/code/CodePractices.java
@@ -6,7 +6,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
-import java.util.regex.Pattern;
 
 public class CodePractices {
 
@@ -108,27 +107,6 @@ public class CodePractices {
     		resultFrequency.computeIfPresent(b.toLowerCase().charAt(i), (k,v) -> v-1);
     	Integer absSum = resultFrequency.values().stream().reduce(0, (x,y) -> Math.abs(x)+Math.abs(y));
     	return absSum == 0;
-    }
-	
-	public static String pangrams(String s) {
-    	final String letters = "abcdefghijklmnopqrstuvxywz";
-    	char[] charArr = letters.toCharArray();
-    	int[] qt = new int[charArr.length];    	
-    	if (s == null || s.length() < charArr.length)
-    		return "not pangram";    	
-    	for (int i=0; i<s.length(); i++) {
-    		char currChar = s.toLowerCase().charAt(i);
-    		for(int j=0; j<charArr.length; j++) {
-    			if (currChar == charArr[j]) {
-    				qt[j]++;
-    				break;
-    			}
-    		}
-    	}    	
-    	boolean result = true;
-    	for(int j=0; j<charArr.length; j++) 
-    		result = result && (qt[j] > 0);
-    	return result ? "pangram" : "not pangram";	
     }
 					
 }
